First

+++ Adapter/src/meteo/SensorTemperature.java
@@ -2,7 +2,7 @@ package meteo;
 
 interface SensorTemperature{
 	int identifier(); // идентификатор датчика
-	double temperature(); // температура датчика
+	Float temperature(); // температура датчика
 	int year();			 // Год
 	int day();			 // День года
 	int second();		 // Секунда дня
